ai_server/requirements.txt
fastapi
uvicorn
firebase-admin
paho-mqtt
